requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.7.1
pydantic-settings==2.2.1
redis==5.0.4
httpx==0.27.0
apscheduler==3.10.4
python-dotenv==1.0.1
pytest==8.2.1
pytest-asyncio==0.23.7
pytest-mock==3.14.0
